code/python_client/requirements.txt
numpy
gym
pyyaml
opencv-python
